Add informations about the suggestions database
The file is a gziped SQLite 3 database with the .jpg extension to
prevent AAPT from handling the compression. This way the system doesn't
know that the file is compressed and does not raise an error with
Android versions prior to 2.3 because the uncompressed file size is over
1MB.

The uncompression must be carried out by ourselves in the source code.