Add an AsyncTask to install the database
When executed, this task uncompresses the database from the assets to
the internal storage. A DatabaseInstallationListener can be attached to
it to be notified of the progress and completion of the database
installation. The listener can also be detached and re-attached by an
activity to pass through a configuration change without leaking the
activity.