Add an ExpandableTextView widget
This widget gets the collapsed TextView height and the gradient height
respectively from @dimen/collapsed_textview_height and
@dimen/collapsed_textview_gradient_height.

If the sum of the height of the TextView and the height of the expand
arrow is lower than the collapsed TextView height, then the TextView
will not be collapsed but will remain like a regular TextView.
Otherwise, it will be collapsed and a gradient will be applied to the
text. The gradient is removed when the textview is expanded and
re-applied when it is collapsed again.