SuggestionsProvider: make searching much faster
The SQL query has been rewritten to merely return all the matching forms
without trying to group them by ARK name. The grouping was complicated
because of the form type handling, and SQLite could not handle it in a
reasonnable time frame. The grouping is now done in the Java code, inside
a CursorWrapper. The searching time is now significantly faster.