MainActivity: check disk space for DB installation
Check that there is enough available disk space on the device before installing
the database (size of the uncompressed database + 1 MB for safety). If there is
not enough available disk space, the user is asked to uninstall some
applications. The available disk space is monitored and the database
installation starts automatically when the required disk space is available.